requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow 